use velocity and acceleration in agent (#5)
* use velocity and acceleration in agent

* adjust formatting

* normalize velocity observation

* Update swarm/bas/swarm.py


* remove if observation_space None

* Update swarm/bas/agent.py


* change velocities etc. to floats

* Update swarm/bas/wrappers/section_and_velocity_observation.py


* Update swarm/bas/agent.py


* use binary or

* convert to int

+++ swarm/bas/swarm.py
@@ -12,6 +12,7 @@ def normalize(vectors: np.ndarray, eps: float = 1e-6) -> np.ndarray:
 def limit(vectors: np.ndarray, max_norm: float) -> np.ndarray:
     """Rescale row vectors if their norm is greater than `max_norm`."""
 
+    # Clipping is performed on the lower bound.
     clipped_norms = np.linalg.norm(vectors, axis=-1, keepdims=True).clip(max_norm)
     return vectors * max_norm / clipped_norms
 
@@ -34,9 +35,9 @@ class Swarm:
     def __init__(
         self,
         num_boids: int,
-        radius: int,
-        max_velocity: int | None,
-        max_acceleration: int,
+        radius: float,
+        max_velocity: float | None,
+        max_acceleration: float,
         separation_range: float,
         alignment_range: float,
         cohesion_range: float,
